Delete unused indicatorTitle component
